refactor(UserDeletionUseCase): refactor UserDeletionRepository to UserDeletionUseCase
