+++ app/src/androidTest/java/com/android/unio/components/notification/NotificationTest.kt
@@ -17,7 +17,7 @@ import com.android.unio.model.notification.UnioNotification
 import com.android.unio.model.usecase.SaveUseCaseFirestore
 import com.android.unio.model.search.SearchRepository
 import com.android.unio.model.search.SearchViewModel
-import com.android.unio.model.user.UserDeletionRepositoryFirestore
+import com.android.unio.model.usecase.UserDeletionUseCaseFirestore
 import com.android.unio.model.user.UserRepositoryFirestore
 import com.android.unio.model.user.UserViewModel
 import com.android.unio.ui.home.HomeScreen
@@ -52,7 +52,7 @@ class NotificationTest : TearDown() {
   @MockK private lateinit var imageRepository: ImageRepositoryFirebaseStorage
   @MockK private lateinit var associationRepository: AssociationRepositoryFirestore
   @MockK private lateinit var userRepository: UserRepositoryFirestore
-  @MockK private lateinit var userDeletionRepository: UserDeletionRepositoryFirestore
+  @MockK private lateinit var userDeletionRepository: UserDeletionUseCaseFirestore
   @MockK
   private lateinit var eventUserPictureRepositoryFirestore: EventUserPictureRepositoryFirestore
   @MockK
